Сделал более понятную структуру с помощью модулей
