+++ js/createFormFunctions/createComments.js
@@ -1,0 +1,7 @@
+import {createComment} from './createFormFunctions/createComment.js';
+
+function createComments (cycles) {
+  return Array.from({length: cycles}, createComment);
+}
+
+export {createComments};
