+++ js/createFormFunctions/createWholeForm.js
@@ -1,0 +1,5 @@
+import { createFormObject } from './createFormObject.js';
+
+const similarFormObjects = Array.from({length: 25}, createFormObject);
+
+export {similarFormObjects};
