+++ js/formObjectsPropertiesFunctions/createDescription.js
@@ -1,0 +1,6 @@
+import { getRandomItemFromArray } from './functionsForArray/getRandomItemFromArray.js';
+import { variantsOfDescriptions } from './variables/formVariables.js';
+
+const createDescription = () => getRandomItemFromArray(variantsOfDescriptions);
+
+export {createDescription};
