+++ js/formObjectsPropertiesFunctions/createName.js
@@ -1,0 +1,6 @@
+import { getRandomItemFromArray } from './functionsForArray/getRandomItemFromArray.js';
+import { variantsOfNames } from '../variables/formVariables.js';
+
+const createName = () => getRandomItemFromArray (variantsOfNames);
+
+export {createName};
