+++ js/formObjectsPropertiesFunctions/createUrl.js
@@ -1,0 +1,7 @@
+import {getRandomPositiveInteger} from './mathFunctions/randomNumber.js';
+
+function createUrl() {
+  return `photos/${getRandomPositiveInteger(1,25)}.jpg`;
+}
+
+export {createUrl};
