+++ js/functionsForArray/getRandomItemFromArray.js
@@ -1,0 +1,5 @@
+import { getRandomPositiveInteger } from './mathFunctions/randomNumber.js';
+
+const getRandomItemFromArray = (items) => items[getRandomPositiveInteger(0, items.length-1)];
+
+export {getRandomItemFromArray};
