+++ js/mathFunctions/checkLength.js
@@ -1,0 +1,5 @@
+function checkLength(string, maxLength) {
+  return string.length < maxLength;
+}
+
+export {checkLength};
